slight modification: do a generic batch case, then common ones.
since batch reqs can actually involve an arbitrary mix of APIs,
it becomes difficult to properly manage the types of all that
on the client side.

to simplify, we will have a generic batch case for all types
of batches that mix & match arbitrarily, and then 2 common
ones for IPs-only and ASNs-only which will check the validity
of their inputs.